Revert "Allowing storing GPGK raster layer styles to DB"
