+++ python/core/auto_generated/qgis.sip.in
@@ -2121,15 +2121,6 @@ The development version
     typedef QFlags<Qgis::DatabaseProviderConnectionCapability2> DatabaseProviderConnectionCapabilities2;
 
 
-    enum class ProviderStyleStorageCapability
-    {
-      SaveToDatabase,
-      LoadFromDatabase,
-      DeleteFromDatabase
-    };
-    typedef QFlags<Qgis::ProviderStyleStorageCapability> ProviderStyleStorageCapabilities;
-
-
     enum class UserProfileSelectionPolicy
     {
       LastProfile,
