+++ python/core/auto_generated/vector/qgsvectordataprovider.sip.in
@@ -543,6 +543,18 @@ Clear recorded errors
 Gets recorded errors
 %End
 
+    virtual bool isSaveAndLoadStyleToDatabaseSupported() const;
+%Docstring
+It returns ``False`` by default.
+Must be implemented by providers that support saving and loading styles to db returning ``True``
+%End
+
+    virtual bool isDeleteStyleFromDatabaseSupported() const;
+%Docstring
+It returns ``False`` by default.
+Must be implemented by providers that support delete styles from db returning ``True``
+%End
+
     virtual QgsFeatureRenderer *createRenderer( const QVariantMap &configuration = QVariantMap() ) const /Factory/;
 %Docstring
 Creates a new vector layer feature renderer, using provider backend specific information.
