+++ python/gui/auto_generated/qgslayerpropertiesdialog.sip.in
@@ -27,7 +27,6 @@ Base class for "layer properties" dialogs, containing common utilities for handl
 %End
   public:
 
-
     QgsLayerPropertiesDialog( QgsMapLayer *layer, QgsMapCanvas *canvas, const QString &settingsKey, QWidget *parent /TransferThis/ = 0, Qt::WindowFlags fl = Qt::WindowFlags(), QgsSettings *settings = 0 );
 %Docstring
 Constructor for QgsLayerPropertiesDialog.
@@ -50,27 +49,6 @@ This must be called in order for the standard metadata loading/saving functional
     virtual void addPropertiesPageFactory( const QgsMapLayerConfigWidgetFactory *factory );
 %Docstring
 Adds properties page from a ``factory``.
-%End
-
-    void saveDefaultStyle();
-%Docstring
-Saves the default style when appropriate button is pressed
-
-.. versionadded:: 3.30
-%End
-
-    void loadStyle();
-%Docstring
-Triggers a dialog to load a saved style
-
-.. versionadded:: 3.30
-%End
-
-    void saveStyleAs();
-%Docstring
-Saves a style when appriate button is pressed
-
-.. versionadded:: 3.30
 %End
 
   public slots:
