+++ python/gui/auto_generated/vector/qgsvectorlayerproperties.sip.in
@@ -25,6 +25,34 @@ class QgsVectorLayerProperties : QgsLayerPropertiesDialog
     virtual bool eventFilter( QObject *obj, QEvent *ev );
 
 
+    void loadDefaultStyle();
+%Docstring
+Loads the default style when appropriate button is pressed
+
+.. versionadded:: 3.30
+%End
+
+    void saveDefaultStyle();
+%Docstring
+Saves the default style when appropriate button is pressed
+
+.. versionadded:: 3.30
+%End
+
+    void loadStyle();
+%Docstring
+Loads a saved style when appropriate button is pressed
+
+.. versionadded:: 3.30
+%End
+
+    void saveStyleAs();
+%Docstring
+Saves a style when appriate button is pressed
+
+.. versionadded:: 3.30
+%End
+
   protected slots:
     void optionsStackedWidget_CurrentChanged( int index ) final;
     virtual void syncToLayer() ${SIP_FINAL};
