+++ src/core/providers/qgsdataprovider.cpp
@@ -134,8 +134,3 @@ QString QgsDataProvider::sublayerSeparator()
 {
   return SUBLAYER_SEPARATOR;
 }
-
-Qgis::ProviderStyleStorageCapabilities QgsDataProvider::styleStorageCapabilities() const
-{
-  return Qgis::ProviderStyleStorageCapabilities();
-}
