+++ src/core/providers/qgsdataprovider.h
@@ -633,13 +633,6 @@ class CORE_EXPORT QgsDataProvider : public QObject
      */
     static QString sublayerSeparator();
 
-    /**
-     * Returns the style storage capabilities.
-     *
-     * \since QGIS 3.34
-     */
-    virtual Qgis::ProviderStyleStorageCapabilities styleStorageCapabilities() const;
-
   signals:
 
     /**
