+++ src/gui/CMakeLists.txt
@@ -39,6 +39,7 @@ set(QGIS_GUI_SRCS
   vector/qgssourcefieldsproperties.cpp
   vector/qgsvectorlayerlegendwidget.cpp
   vector/qgsvectorlayerproperties.cpp
+  vector/qgsvectorlayersavestyledialog.cpp
   vector/qgswmsdimensiondialog.cpp
 
   symbology/qgs25drendererwidget.cpp
@@ -621,7 +622,6 @@ set(QGIS_GUI_SRCS
   qgsmaplayerconfigwidgetfactory.cpp
   qgsmaplayerloadstyledialog.cpp
   qgsmaplayerrefreshsettingswidget.cpp
-  qgsmaplayersavestyledialog.cpp
   qgsmaplayerstylecategoriesmodel.cpp
   qgsmaplayerstyleguiutils.cpp
   qgsmaplayerstylemanagerwidget.cpp
@@ -897,7 +897,6 @@ set(QGIS_GUI_HDRS
   qgsmaplayercombobox.h
   qgsmaplayerconfigwidget.h
   qgsmaplayerconfigwidgetfactory.h
-  qgsmaplayersavestyledialog.h
   qgsmaplayerloadstyledialog.h
   qgsmaplayerrefreshsettingswidget.h
   qgsmaplayerstylecategoriesmodel.h
@@ -1432,6 +1431,7 @@ set(QGIS_GUI_HDRS
   vector/qgssourcefieldsproperties.h
   vector/qgsvectorlayerlegendwidget.h
   vector/qgsvectorlayerproperties.h
+  vector/qgsvectorlayersavestyledialog.h
   vector/qgswmsdimensiondialog.h
 
   symbology/characterwidget.h
