+++ src/gui/mesh/qgsmeshlayerproperties.h
@@ -24,6 +24,7 @@
 
 class QgsMapLayer;
 class QgsMapCanvas;
+class QgsMeshLayer;
 class QgsRendererMeshPropertiesWidget;
 class QgsMeshLayer3DRendererWidget;
 class QgsMeshStaticDatasetWidget;
