+++ src/gui/qgsmaplayerloadstyledialog.cpp
@@ -19,7 +19,7 @@
 #include "qgsmaplayerloadstyledialog.h"
 #include "qgslogger.h"
 #include "qgssettings.h"
-#include "qgslayerpropertiesdialog.h"
+#include "qgsvectorlayerproperties.h"
 #include "qgsmaplayerstylecategoriesmodel.h"
 #include "qgshelp.h"
 #include "qgsapplication.h"
@@ -43,29 +43,50 @@ QgsMapLayerLoadStyleDialog::QgsMapLayerLoadStyleDialog( QgsMapLayer *layer, QWid
 
   QgsSettings settings;
 
+  QString providerName = mLayer->providerType();
+  if ( providerName == QLatin1String( "ogr" ) )
+  {
+    QgsVectorLayer *vl = qobject_cast< QgsVectorLayer * >( mLayer );
+    providerName = vl->dataProvider()->storageType();
+    if ( providerName == QLatin1String( "GPKG" ) )
+      providerName = QStringLiteral( "GeoPackage" );
+  }
+
   const QString myLastUsedDir = settings.value( QStringLiteral( "style/lastStyleDir" ), QDir::homePath() ).toString();
 
   // load style type combobox
   connect( mStyleTypeComboBox, qOverload<int>( &QComboBox::currentIndexChanged ), this, [ = ]( int )
   {
-    const QgsLayerPropertiesDialog::StyleType type = currentStyleType();
-    mFileLabel->setVisible( type != QgsLayerPropertiesDialog::StyleType::DatasourceDatabase && type != QgsLayerPropertiesDialog::StyleType::UserDatabase );
-    mFileWidget->setVisible( type != QgsLayerPropertiesDialog::StyleType::DatasourceDatabase && type != QgsLayerPropertiesDialog::StyleType::UserDatabase );
-    mFromDbWidget->setVisible( type == QgsLayerPropertiesDialog::StyleType::DatasourceDatabase );
-    mDeleteButton->setVisible( type == QgsLayerPropertiesDialog::StyleType::DatasourceDatabase && mLayer->dataProvider()->styleStorageCapabilities().testFlag( Qgis::ProviderStyleStorageCapability::DeleteFromDatabase ) );
+    const QgsVectorLayerProperties::StyleType type = currentStyleType();
+    QgsVectorLayer *vl = qobject_cast< QgsVectorLayer * >( mLayer );
+    mFileLabel->setVisible( !vl || ( type != QgsVectorLayerProperties::StyleType::DB && type != QgsVectorLayerProperties::StyleType::Local ) );
+    mFileWidget->setVisible( !vl || ( type != QgsVectorLayerProperties::StyleType::DB && type != QgsVectorLayerProperties::StyleType::Local ) );
+    if ( vl )
+    {
+      mFromDbWidget->setVisible( type == QgsVectorLayerProperties::StyleType::DB );
+      mDeleteButton->setVisible( type == QgsVectorLayerProperties::StyleType::DB && vl->dataProvider()->isDeleteStyleFromDatabaseSupported() );
+    }
+    else
+    {
+      mFromDbWidget->setVisible( false );
+      mDeleteButton->setVisible( false );
+    }
 
-    mStyleCategoriesListView->setEnabled( currentStyleType() != QgsLayerPropertiesDialog::StyleType::SLD );
+    mStyleCategoriesListView->setEnabled( !vl || currentStyleType() != QgsVectorLayerProperties::StyleType::SLD );
     updateLoadButtonState();
   } );
-  mStyleTypeComboBox->addItem( tr( "From file" ), QgsLayerPropertiesDialog::QML ); // QML is used as entry, but works for SLD too, see currentStyleType()
-  mStyleTypeComboBox->addItem( tr( "Default from local database" ), QgsLayerPropertiesDialog::UserDatabase );
+  mStyleTypeComboBox->addItem( tr( "From File" ), QgsVectorLayerProperties::QML ); // QML is used as entry, but works for SLD too, see currentStyleType()
+  mStyleTypeComboBox->addItem( tr( "Default from local database" ), QgsVectorLayerProperties::Local );
 
-  if ( mLayer->dataProvider()->styleStorageCapabilities().testFlag( Qgis::ProviderStyleStorageCapability::LoadFromDatabase ) )
+  if ( QgsVectorLayer *vl = qobject_cast< QgsVectorLayer * >( mLayer ) )
   {
-    mStyleTypeComboBox->addItem( tr( "From datasource database" ), QgsLayerPropertiesDialog::StyleType::DatasourceDatabase );
-    if ( settings.value( QStringLiteral( "style/lastLoadStyleTypeSelection" ) ) == QgsLayerPropertiesDialog::StyleType::DatasourceDatabase )
+    if ( vl->dataProvider()->isSaveAndLoadStyleToDatabaseSupported() )
     {
-      mStyleTypeComboBox->setCurrentIndex( mStyleTypeComboBox->findData( QgsLayerPropertiesDialog::StyleType::DatasourceDatabase ) );
+      mStyleTypeComboBox->addItem( tr( "From Database (%1)" ).arg( providerName ), QgsVectorLayerProperties::StyleType::DB );
+      if ( settings.value( QStringLiteral( "style/lastLoadStyleTypeSelection" ) ) == QgsVectorLayerProperties::StyleType::DB )
+      {
+        mStyleTypeComboBox->setCurrentIndex( mStyleTypeComboBox->findData( QgsVectorLayerProperties::StyleType::DB ) );
+      }
     }
   }
 
@@ -103,7 +124,8 @@ QgsMapLayerLoadStyleDialog::QgsMapLayerLoadStyleDialog( QgsMapLayer *layer, QWid
   mFileWidget->setDefaultRoot( myLastUsedDir );
   connect( mFileWidget, &QgsFileWidget::fileChanged, this, [ = ]( const QString & path )
   {
-    mStyleCategoriesListView->setEnabled( currentStyleType() != QgsLayerPropertiesDialog::SLD );
+    QgsVectorLayer *vl = qobject_cast< QgsVectorLayer * >( mLayer );
+    mStyleCategoriesListView->setEnabled( !vl || currentStyleType() != QgsVectorLayerProperties::SLD );
     QgsSettings settings;
     const QFileInfo tmplFileInfo( path );
     settings.setValue( QStringLiteral( "style/lastStyleDir" ), tmplFileInfo.absolutePath() );
@@ -145,14 +167,14 @@ QgsMapLayer::StyleCategories QgsMapLayerLoadStyleDialog::styleCategories() const
   return mModel->categories();
 }
 
-QgsLayerPropertiesDialog::StyleType QgsMapLayerLoadStyleDialog::currentStyleType() const
+QgsVectorLayerProperties::StyleType QgsMapLayerLoadStyleDialog::currentStyleType() const
 {
-  QgsLayerPropertiesDialog::StyleType type = mStyleTypeComboBox->currentData().value<QgsLayerPropertiesDialog::StyleType>();
-  if ( type == QgsLayerPropertiesDialog::QML )
+  QgsVectorLayerProperties::StyleType type = mStyleTypeComboBox->currentData().value<QgsVectorLayerProperties::StyleType>();
+  if ( type == QgsVectorLayerProperties::QML )
   {
     const QFileInfo fi( mFileWidget->filePath() );
     if ( fi.exists() && fi.suffix().compare( QStringLiteral( "sld" ), Qt::CaseInsensitive ) == 0 )
-      type = QgsLayerPropertiesDialog::SLD;
+      type = QgsVectorLayerProperties::SLD;
   }
   return type;
 }
@@ -276,6 +298,10 @@ void QgsMapLayerLoadStyleDialog::accept()
 
 void QgsMapLayerLoadStyleDialog::deleteStyleFromDB()
 {
+  QgsVectorLayer *vl = qobject_cast< QgsVectorLayer *>( mLayer );
+  if ( !vl )
+    return;
+
   QString msgError;
   const QString opInfo = QObject::tr( "Delete style %1 from %2" ).arg( mSelectedStyleName, mLayer->providerType() );
 
@@ -284,7 +310,7 @@ void QgsMapLayerLoadStyleDialog::deleteStyleFromDB()
                               QMessageBox::Yes | QMessageBox::No, QMessageBox::No ) != QMessageBox::Yes )
     return;
 
-  mLayer->deleteStyleFromDatabase( mSelectedStyleId, msgError );
+  vl->deleteStyleFromDatabase( mSelectedStyleId, msgError );
   if ( !msgError.isNull() )
   {
     QgsDebugError( opInfo + " failed." );
@@ -302,7 +328,7 @@ void QgsMapLayerLoadStyleDialog::deleteStyleFromDB()
     QString errorMsg;
     QStringList ids, names, descriptions;
     //get the list of styles in the db
-    const int sectionLimit = mLayer->listStylesInDatabase( ids, names, descriptions, errorMsg );
+    const int sectionLimit = vl->listStylesInDatabase( ids, names, descriptions, errorMsg );
     if ( !errorMsg.isNull() )
     {
       QMessageBox::warning( this, tr( "Error occurred while retrieving styles from database" ), errorMsg );
@@ -316,12 +342,19 @@ void QgsMapLayerLoadStyleDialog::deleteStyleFromDB()
 
 void QgsMapLayerLoadStyleDialog::updateLoadButtonState()
 {
-  const QgsLayerPropertiesDialog::StyleType type = currentStyleType();
-  mLoadButton->setEnabled( ( type == QgsLayerPropertiesDialog::DatasourceDatabase
-                             && ( mRelatedTable->selectionModel()->hasSelection() || mOthersTable->selectionModel()->hasSelection()
-                                ) ) ||
-                           ( type != QgsLayerPropertiesDialog::DatasourceDatabase && !mFileWidget->filePath().isEmpty() ) ||
-                           type == QgsLayerPropertiesDialog::UserDatabase );
+  const QgsVectorLayerProperties::StyleType type = currentStyleType();
+  if ( mLayer->type() == Qgis::LayerType::Vector )
+  {
+    mLoadButton->setEnabled( ( type == QgsVectorLayerProperties::DB
+                               && ( mRelatedTable->selectionModel()->hasSelection() || mOthersTable->selectionModel()->hasSelection()
+                                  ) ) ||
+                             ( type != QgsVectorLayerProperties::DB && !mFileWidget->filePath().isEmpty() ) ||
+                             type == QgsVectorLayerProperties::Local );
+  }
+  else
+  {
+    mLoadButton->setEnabled( !mFileWidget->filePath().isEmpty() );
+  }
 }
 
 void QgsMapLayerLoadStyleDialog::showHelp()
