+++ src/gui/qgsmaplayerstylecategoriesmodel.cpp
@@ -30,8 +30,6 @@ QgsMapLayerStyleCategoriesModel::QgsMapLayerStyleCategoriesModel( Qgis::LayerTyp
       break;
 
     case Qgis::LayerType::Raster:
-      mCategoryList << QgsMapLayer::StyleCategory::Symbology << QgsMapLayer::StyleCategory::AllStyleCategories;
-      break;
     case Qgis::LayerType::Annotation:
     case Qgis::LayerType::Plugin:
     case Qgis::LayerType::Mesh:
@@ -43,11 +41,7 @@ QgsMapLayerStyleCategoriesModel::QgsMapLayerStyleCategoriesModel( Qgis::LayerTyp
   }
 
   // move All categories to top
-  int idxAllStyleCategories = mCategoryList.indexOf( QgsMapLayer::AllStyleCategories );
-  if ( idxAllStyleCategories > 0 )
-  {
-    mCategoryList.move( idxAllStyleCategories, 0 );
-  }
+  mCategoryList.move( mCategoryList.indexOf( QgsMapLayer::AllStyleCategories ), 0 );
 }
 
 void QgsMapLayerStyleCategoriesModel::setCategories( QgsMapLayer::StyleCategories categories )
@@ -75,7 +69,7 @@ void QgsMapLayerStyleCategoriesModel::setShowAllCategories( bool showAll )
 int QgsMapLayerStyleCategoriesModel::rowCount( const QModelIndex & ) const
 {
   int count = mCategoryList.count();
-  if ( count > 0 && !mShowAllCategories )
+  if ( !mShowAllCategories )
     count--;
   return count;
 }
