+++ src/providers/oracle/qgsoracleprovider.h
@@ -185,7 +185,7 @@ class QgsOracleProvider final: public QgsVectorDataProvider
 
     static bool execLoggedStatic( QSqlQuery &qry, const QString &sql, const QVariantList &args, const QString &uri, const QString &originatorClass = QString(), const QString &queryOrigin = QString() );
 
-    Qgis::ProviderStyleStorageCapabilities styleStorageCapabilities() const override;
+    bool isSaveAndLoadStyleToDatabaseSupported() const override { return true; }
     void setTransaction( QgsTransaction *transaction ) override;
     QgsTransaction *transaction() const override;
