+++ src/ui/qgsvectorlayersavestyledialog.ui
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <ui version="4.0">
- <class>QgsMapLayerSaveStyleDialog</class>
- <widget class="QDialog" name="QgsMapLayerSaveStyleDialog">
+ <class>QgsVectorLayerSaveStyleDialog</class>
+ <widget class="QDialog" name="QgsVectorLayerSaveStyleDialog">
   <property name="geometry">
    <rect>
     <x>0</x>
@@ -200,7 +200,7 @@
   <connection>
    <sender>buttonBox</sender>
    <signal>accepted()</signal>
-   <receiver>QgsMapLayerSaveStyleDialog</receiver>
+   <receiver>QgsVectorLayerSaveStyleDialog</receiver>
    <slot>accept()</slot>
    <hints>
     <hint type="sourcelabel">
@@ -216,7 +216,7 @@
   <connection>
    <sender>buttonBox</sender>
    <signal>rejected()</signal>
-   <receiver>QgsMapLayerSaveStyleDialog</receiver>
+   <receiver>QgsVectorLayerSaveStyleDialog</receiver>
    <slot>reject()</slot>
    <hints>
     <hint type="sourcelabel">
