+++ tests/src/python/test_provider_mssql.py
@@ -15,7 +15,6 @@ import qgis  # NOQA
 from qgis.PyQt.QtCore import QDate, QDateTime, QDir, QTime, QVariant
 from qgis.core import (
     NULL,
-    Qgis,
     QgsCoordinateReferenceSystem,
     QgsDataProvider,
     QgsDataSourceUri,
@@ -404,8 +403,8 @@ class TestPyQgsMssqlProvider(QgisTestCase, ProviderTestCase):
 
         vl = self.getSource()
         self.assertTrue(vl.isValid())
-        self.assertEqual(int(vl.dataProvider().styleStorageCapabilities()) & Qgis.ProviderStyleStorageCapability.LoadFromDatabase, Qgis.ProviderStyleStorageCapability.LoadFromDatabase)
-        self.assertEqual(int(vl.dataProvider().styleStorageCapabilities()) & Qgis.ProviderStyleStorageCapability.SaveToDatabase, Qgis.ProviderStyleStorageCapability.SaveToDatabase)
+        self.assertTrue(
+            vl.dataProvider().isSaveAndLoadStyleToDatabaseSupported())
 
         # table layer_styles does not exist
